Functions moved to utils

+++ Extrapolation/Extrapolation/exp_smoothing.h
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <vector>
 
-std::vector<double> exponential_smoothing(std::vector<double> x, std::vector<double> y, std::vector<double> x1, float alpha);
+std::vector<double> exponential_smoothing(std::vector<double> x, std::vector<double> y, std::vector<double> x1, double alpha);
